classpulse/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.40.0
python-dotenv>=1.0.0
